+++ opendut-carl/src/persistence/resources/cluster_configuration.rs
@@ -8,11 +8,6 @@ impl Persistable for ClusterConfiguration {
     fn insert(self, _id: ClusterId, storage: &mut Storage) -> PersistenceResult<()> {
         let mut connection = storage.db.connection();
 
-        //Delete before inserting to ensure that when an update removes
-        //list elements we don't leave those elements behind in the database.
-        //TODO more efficient solution
-        query::cluster_configuration::remove(self.id, &mut connection)?;
-
         query::cluster_configuration::insert(self, &mut connection)
     }